Commissions form
Linked commissions form to commissions link, if user is signed in but has no commissions a smart link directs them to the form, if they have quotes it directs them to the dashboard. If user clicks commissions, it asks them to sign in and then follows the same smart link.